live

+++ src/pages/logout.js
@@ -2,11 +2,17 @@ import React, { useContext } from "react"
 import { store } from "src/store"
 import SEO from "src/components/seo"
 import { navigate } from "gatsby"
+import useIsClient from "src/hooks/use-is-client"
 
 const Logout = () => {
+  const { isClient, key } = useIsClient()
   const sessionMachine = useContext(store)
   const { state, send } = sessionMachine
-  window.localStorage.clear()
+  if (!isClient) return null
+  console.log(window)
+  if (window !== undefined) {
+    window.localStorage.clear()
+  }
   console.log("sending LOGOUT", state.value)
   send("LOGOUT")
   setTimeout(() => {
@@ -15,8 +21,10 @@ const Logout = () => {
   return (
     <>
       <SEO title="Logout" />
-      <h1>Goodbye!</h1>
-      <p>See you next time.</p>
+      <div className="p-5">
+        <h1>Goodbye!</h1>
+        <p>See you next time.</p>
+      </div>
     </>
   )
 }
